refactor: Splitted request helpers to one file per class
