+++ src/stream/TweetStream.ts
@@ -1,6 +1,7 @@
 import { EventEmitter } from 'events';
 import type { IncomingMessage, ClientRequest } from 'http';
-import { RequestHandlerHelper, TRequestFullData } from '../client-mixins/request-maker.mixin';
+import RequestHandlerHelper from '../client-mixins/request-handler.helper';
+import { TRequestFullData } from '../client-mixins/request-maker.mixin';
 import { ETwitterStreamEvent } from '../types';
 import TweetStreamEventCombiner from './TweetStreamEventCombiner';
 import TweetStreamParser, { EStreamParserEvent } from './TweetStreamParser';
